refactor: cleanup component locations
