+++ packages/desktop/views/dashboard/components/Sidebar.svelte
