+++ packages/desktop/views/dashboard/components/index.js
@@ -1,0 +1,2 @@
+export { default as Navbar } from './Navbar.svelte'
+export { default as Sidebar } from './Sidebar.svelte'
